refact: reorganize files for redux-form service
